qRestAPIConfig - Do not explicitly find library - Rely on exported target
Since target are exported, it is not needed to use "find_library" to
get the full path of the library. After calling "find_package(qRestAPI)",
project can safely use "${qRestAPI_LIBRARIES}" or "qRestAPI" within the
"target_link_libraries" call.

Not also that since the library is imported
as a CMake target, using "target_link_directories" is also obsolete.