Support windows static build
Following commit 5c056f2, the project could be built as a shared library
this caused problem on windows where the qRestAPI_EXPORT macro was
also defined in the case of static build causing an "inconsistent linkage"
error.

This commit explicitly define qRestAPI_EXPORT based on the
BUILD_SHARED_LIBS option.