Fix build CLang build error by making static helper method public.
For the record, the reason this fails on Mac and not on Linux is
the following :
http://clang-developers.42468.n3.nabble.com/Access-to-static-protected-member-from-friends-of-derived-classes-td4030444.html

The Q_DECLARE_PRIVATE macro declares the qXnatAPIPrivate class and makes
it a friend of qXnatAPI, which is sufficient for G++ to give it access to
qRestAPI's static protected methods, but the Clang developers have a
different understanding of the standard, so it doesn't work on Clang.

Thanks to Steve Pieper, Florian Vichot and Sascha Zelzer for their
help fixing the issue.

See #334