Deleting the QNetworkManager instance too early may lead to crashes.
The qRestAPIPrivate::NetworkManager member was created with a
qRestAPI* parent which lead to the immediate destruction of the
QNetworkManager instance when the qRestAPI instance was deleted.
In turn, this lead to network errors and subsequent crashes in
some scenarios where QNetworkReply instances where still alive.
Not putting qRestAPIPrivate::NetworkManager in a QObject hierarchy
and explicitly calling "deleteLater()" on it solves this issue.