BUG: correctly forward errors
This is really strange; qRestAPI::takeResult() returns a qRestResult pointer, which contains error code+string, yet if there is an error, that information is copied into qRestAPI’s members and NULL is returned instead.  Therefore, this code never saw an error.  Instead, we have to query restAPI.error() after the fact.